+++ lib/rubocop/cop/style/leading_comment_space.rb
@@ -17,6 +17,11 @@ module RuboCop
             next unless comment.text =~ /\A#+[^#\s=:+-]/
             next if comment.text.start_with?('#!') && comment.loc.line == 1
 
+            # in config.ru files, if the first line starts with #\ it is treated
+            # as options (e.g. `#\ -p 8765` sets the request port to 8765)
+            next if comment.text.start_with?('#\\') && comment.loc.line == 1 &&
+                    config_ru?(processed_source.buffer.name)
+
             add_offense(comment, :expression)
           end
         end
@@ -26,6 +31,12 @@ module RuboCop
           b = expr.begin_pos
           hash_mark = range_between(b, b + 1)
           ->(corrector) { corrector.insert_after(hash_mark, ' ') }
+        end
+
+        private
+
+        def config_ru?(file_path)
+          File.basename(file_path).eql?('config.ru')
         end
       end
     end